Change label.
This is a syntax only change for how the label is defined for SP800108. The XAES spec document originally defined it as the character 'X'. The spec changed the notation to be more clear that it is a single octet with the value 0x58.

This changes the implementation to follow the notation of XAES.